app/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
transformers==4.41.1
datasets==2.19.1
torch==2.3.0
accelerate==0.30.1
pydantic==2.7.1
